refactored server and data
